refactor: rename syntax package into core
